+++ MGBoxKit/MGTableBoxStyled.h
@@ -8,11 +8,14 @@
 /**
 * `MGTableBoxStyled` is a thin wrapper of MGTableBox, providing default size,
 * background colour, shadow, and corner rounding. The styling is designed to
-* provide a table section style similar to `UITableView` grouped style.
+* provide a table section style similar to pre iOS 7 `UITableView` grouped style.
 * It also serves as a basic subclassing example (take a look at
 * `MGTableBoxStyled.m` for example `setup` and `layout` methods).
+*
+* @deprecated The superclass is no longer maintained. This class should only be
+* used as a loose example of subclasssing.
 */
 
-@interface MGTableBoxStyled : MGTableBox
+__deprecated @interface MGTableBoxStyled : MGTableBox
 
 @end